Loads of changes!
i) Added a new check to see whether any clones have been assigned to more than one Sequence
Map.  Because of this I added a new class of errors ('Clone error -') for the subroutine that
generates HTML, I suggest we use this for any other clone checks we might implement.

ii) updated $tace to point to 4.9c version of tace

iii) changed the version of babel.pl that was being used.  Now points to the /wormsrv2/scripts
and not to the ~wormpub/analysis/scripts copy

iv) Fixed (hopefully) an older test of whether predicted genes belong to SUPERLINK objects, a
typo/coding error meant that this test didn't actually work

v) Removed the existing get_script call (which used a function in babel.pl to find the version
number of a script) and replaced with a nifty cvs commmand line call (with some parsing).

In general I think we should use this if we want any scripts to fetch/handle version numbers.
The CVS system allows for less ambiguity (every change gets a new version number).