Much tidying up of camcheck.
camcheck is now run with the -w option.  This initially generated many errors,
all of which I have cleared up.  The 'require babel.pl' is no longer needed as
this was just grabbing an (incorrect) version number so I have removed this.

Added '|| die(...)' statements to all filehandle 'open' commands.