Bye bye files.
All of these are the last of those scripts that can be *safely* removed, i.e. where the latest
version of a script (e.g. 'aceprep_4.0') is the same as the basic, unsuffixed version
(e.g. 'aceprep')

Many more scripts still to be removed, but these will require some inspection.