New script: unpack_db.pl
This will hopefully replace the need for having separate scripts to unpack the
briggsae, cshace, and stlace data.  It can unpack each one separately, or you
can ask it to try to do all three.

Usage:

unpack_db.pl -b <date> -c <date> -s <date>

Where -b, -c, and -s correspond to brigace, cshace, and stlace and <date>
is the six figure date format corresponding to the date specified in
the *.tar.gz files for each database.

Can also just run -h for more details.

I suggest that at the next rebuild we try using it safely by only specifying
one database at a time.  If that seems to work then we can proceed (at the
subsequent rebuild) to see if it can cope with doing all three in one go.
I've tested this as much as I can, but I'll only really know if it is
behaving when we try it on the real data.

One other thing, in the unpack_briggsae script I noticed that it would sometime
refer to the 'briggsae' data and sometimes to 'brigace' data.  I've made this
script systematically refer to the data as 'brigace' to be consistent.

This script currently writes to three separate log files (as before) and will mail
all of us a logfile for each database.  If this works well in the future we may
wish to consider changing to writing everything to one single unpack_db log
file.

Oh, and this is a perl5.6.0 script!



New module: Wormbase.pm

I have taken the babel.pl script and turned it into a fully fledged perl module.
I have kept everything that was already there but additionally added a subroutine
to grab the cvs version number.  E.g. just add the following to your script
to get the version number:

my $version = &get_cvs_version($0); # where $0 is the perl variable containing
the full script name.

To use this module you *must* include both of the following in your scripts:

use lib "/wormsrv2/scripts/";	# helps locate the module
use Wormbase;

See the perldoc page for more info...there were a few undocumented subroutines
that were in babel.pl (e.g. tace, gff_sort).  These have all been moved to
Wormbase.pm, so nothing has been lost.  The only change I have made is that
the get_script_version subroutine is not exported by default (we should not
really be using this now that we have get_cvs_version instead).