Have changed the routine for grabbing the current release version number of Wormbase. The routine now grabs the number from the file /wormsrv2/autoace/wspec/database.wrm
In future this file will be changed at the START of any rebuild as the very first step
so if we have started a rebuild then any script requiring the latest version number
will get the correct version number.

Note that there (currently) is a trailing space character in the returned string
character.  This is present in the database.wrm file