The -b (build) flag in autoace_minder would previously be the same as specifying -123.  Now it is only the same as specifying -12.  The -3 option can still be run independently.
The reason for this is because the -3 flag (which creates the distribution files)
should not be run until all the necessary data is in autoace.  In the past this would
be ok because the -12 options would put all the necessary data in autoace.  This is
no longer the case, so you need to run 'autoace_minder -3' at the end of the rebuild.