One tiny error from running make_wormpep was that it couldn't perform chmod on the wp.log59 file.
This is because there is no wp.log59 file! This in turn is because the line in the
script that did create this file is commented out.  Presumably because there is nothing
in the script that seems to write to this log file.

For now I jave just commented out the offending 'chmod' line but we can probably remove
these lines altogether (which would also mean removing error subroutine number 4).

Keith