Take two bottles (and several hundred perl scripts) into the shower?
Not me, I just want to build 'n' go.

Have now automated a lot of the fiddly bits at the end of the rebuild.  Previously
the following stages were run by hand after running make_FTP_public:


  rm ~ftp/pub/wormbase/WSxx/picture.tar.gz
  mkdir ~ftp/pub/wormbase/WSxx/pictures
  cp ~sylvia/geneace/pictures/* ~ftp/pub/wormbase/WSxx/
  cp /wormsrv2/autoace/COMPARE/WSold-WSnew.dbcomp ~ftp/pub/wormbase/WSnew
  tar_wormpep xx
  cp /wormsrv2/WORMPEP/wormpepxx/wormpepxx.tar.gz /nfs/disk69/ftp/pub/wormbase/WSnew


These have now all been incorporated into the make_FTP_public script...with the exception
of the first 'rm ...' command.  To avoid removing this file, I have commented out the
part of the script that was creating it in the first place.

Thunderbirds are go!

FAB.