New (and fairly trivial) function for Wormbase.pm
get_wormbase_release_date();

E.g. $date = &get_wormbase_release_date;

Where $date would end up with a string such as "21st September".

The date in question is taken from the timestamp of the letter.WSxx file in
the wormbase ftp directory.  I could have used other files or other approaches
to get a date, but I figure that this file is only created right at the end
of the rebuild.

The reason for going to these lengths to have such a function is that some of
the Wormbase web pages include the date of the release and this is another thing
that is currently done by hand.

The function does handle the day and month separately so it could be possible
to modify it to just extract either one.

Did you know that there are only 7 days in a (31 day) month that don't need
a 'th' suffix?  Well you do now.