Rejoice for the good work that you have done in your life has been rewarded with...
update_website.pl

This script will change your life...forever. Well maybe.

This monster script does the following things (note that most of these have presently
been done by hand at the end of the rebuild and maybe only Dan truly knew how to
do all these things):

1) Runs copy2web - more specifically I've rewritten a lot of copy2web and integrated it
directly into this new script, so I will probably remove copy2web in the near future.

2) copies the EST_analysis.html file from the previous release and updates the totals
within this file

3) Does the equivalent of what 'mk_EST_pages' was doing by making lots of EST pages

4) Creates all the main (top level) web pages for WORMBASE with updated information
about the new release (this includes the wormpep page under the WORMBASE pages)

5) Updates the wormpep pages (separate from the page mentioned above).  I have gone
back and filled in lots of missing data on the downloads page (this listed the size of
previous wormpep versions).  I have amended this to all work properly (the existing
script to do this had a few holes in).  I've also removed the 'size in Mb' column
from the webpage as this seemed a little redundant.  The wormpep download page is at:

http://www.sanger.ac.uk/Projects/C_elegans/wormpep/wormpep_download.shtml

It now stores the exact date of the wormpep release (rather than just the month).

6) The transfer_all script that previously lurked in the bin/ directory on the web site
has been integrated into the script to copy all sort of GFF files and process them through
gff_sort.

7) All relevant numbers on the data consistency checks page are updated automatically.  I
don't think this always happened in some of the recent releases.

8) This script now launches the list_loci_designations script which we currently run from
within post_build_checks.pl.  Since it is not really a check, I thought I'd move this script
into my new one.  I will later remove it as an option from post_build_checks.


That's about it.  Basically this script should do everything that needs to be done with the website
and should be the last part of the build.

Any future changes to website pages which need to be changed with each release
should be done through this script.