I have added a new check routine to this script.  Normally this script is called by make_wormpep to produce a patch file to read into pepace to make sure pepace is current with the latest wormpep release.
If a CDS is changed it would normally get a new wormpep number and a line
'Wormpep Inactive Replaced_by WP:CExxxxx' would be added to the old Wormpep protein.

However for just a few proteins a problem occurs where there initially are more than
one CDS that correspond to the Wormpep protein.  E.g. Sequence A and Sequence B have
the same encoded protein sequence and would therefore share the same Wormpep protein.

If Sequence A is then changed at some point, it will get a new Wormpep ID but the old
Wormpep ID gets the Inactive tag, even though it still has a valid CDS.

The changes I have made to the script checks to see if the existing Wormpep protein
has more than one CDS in pepace and if it does then it omits to print the 'Replaced_by'
line.

This is probably not the best solution as I guess if both Sequence A and Sequence B are
changed within the same release, then the Replaced_by tag would be needed.  THe best
solution will probably require a change to the Protein model.

For now this problem only affects 8 Wormpep proteins which have an Inactive tag which they
don't need.  I will fix these by hand by removing the Inactive and Replaced_by tags.