New script!
Automates the first bit of the rebuild, i.e.

Looks to see if there is a WSxx release in /wormsrv2 which is three releases old.  If
WS56 was the last build then it will know that WS53 needs to be archived away to make space
for the new WS57 build.

If this folder exists it removes unnecessary directories and creates a *.tar.gz file,
moves it into /wormsrv2/wormbase_archive and deletes the original folder.

It then does a similar thing for an old wormpep release, e.g. /wormsrv2/WORMPEP/wormpep63.

So we shouldn't have to worry about removing any more old releases from /wormsrv2.
Using this script in the rebuild means that we will be storing the last three releases
in uncompressed form.  There is also the current build which is accessible which means
that there is about a months wormbase data available at any given time.

Usage is just 'archive_dbs.pl'