Need to remove homology objects with no coordinates (that is Pep_homol, DNA_homol, and Motif_homol) from the dumped Sequence class of each primary database.
Added the following pattern to the subroutine that strips lines:

/(Pep|DNA|Motif)_homol\s+"[\d:+\w-\.]+"$/

This should match only those homologies with no sets of coordinates at the end (coordinates
don't get enclosed in "" when dumped as an ace file).

This strip routine now runs several times for each pattern which isn't very efficient
(i.e. it scans the dumped file for EST_GENOME and then for "WP:" and then for the above.
Could be improved at some point

Keith