Added a new subroutine 'get_wormrna_version' to get the version number of the latest WormRNA release (currently '1').  It does this by looking at the file
/wormsrv2/WORMRNA/current_WormRNA/README.*

The '*' at the end of the README file should be 'wormrnax' where x is the number.

The 'current_WormRNA' part of the above path is a symbolic link.  This is all necessary
because I guess there won't always be a new release of WormRNA with every new build
of WormBase so we can't just subtract a fixed number from the WS release number like
we do with WormPep.