A new script to do (just about) everything at the start of the build with no pesky human interference.
We currently have to do this at the start of the build:

- archive_dbs.pl  (tidies away old version of WSxx and wormpepxx)

- TransferDB -start /wormsrv2/autoace -end /wormsrv2/WSxx -all -name WSxx

- TransferDB -start /wormsrv1/camace -end /wormsrv2/camace -all -name camace

- modify /wormsrv2/autoace/wspec/database.wrm to new release number (use cvs!)

- recreate symbolic link /wormsrv2/current_DB to point to last release


All of this is now done with prepare_build.pl.  You will still have to remove
the TransferDB logs after this script runs (this scrips just calls the TransferDB
script twice).

If this all works and is stable, then we could even try including the 'unpack_dbs.pl'
script into it, though you would need to get that script to identify what are the latest
files on the FTP site which need unpacking.

So (hopefully) the first part of the build is now all done by just two scripts.

I bet this won't work first time though.

Keith