I have removed the -p flag from make_acefiles which previously produced a peptide dataset.  This is because we will no longer generate peptides at this stage. Instead, peptides will only be added for Protein objects that have wublastx or wublastp homologies (courtesy of Marc's pipeline).
This should ensure that in each build we only have Protein objects (and corresponding
peptide, species info) which are actually linked to something that people can see.
Currently many of our proteins are not linked to anything, or do not have a peptide
sequence, or only have a peptide sequence and nothing else.

The whole produce peptides subroutine has been removed from this script.

Keith