Clone class ----------- Added UNIQUE to Clone tag Added Fosmid as a new Clone type
Sequence class
--------------
Made Analysis_history UNIQUE (i.e. only keep details of last BLAST analysis)
Removed two redundant tags beneath this (parent & child)

Protein class
-------------
Removed Inactive tag under Wormpep heading and replaced with 'Live' but the 'Live' tag
is now on a new line and not buried in with the 'Replaced_by' tag.

So in future all proteins that have ever been in Wormpep will have a 'Wormpep' tag (which
isn't the case at the moment) but only those that correspond to the current Wormpep release will
have a 'Live' tag.  Presence of a 'Live' tag rather than absence of an 'Inactive' tag is
a) more logical and b) more consistant with the use of the 'Live' tag in the new Gene model.

Keith