Added a 'CGC_approved' tag to the existing Locus and Gene_class models.
This is primarily for me to keep track of the data I get from Jonathan that is
rubber stamped with CGC approval. These tags might not necessarily exist in the
new Gene models as I'm hoping that Locus objects in with 'CGC_approved' tags will
simply use the CGC_name field in the new Gene model.