Expanded Gene_class model to allow better, cleaner XREFinf with Gene class
Removed use of WTP_name

Tidied up history section of Gene class (specifically all of the Merged, Split, Aquires
tags) to make it consistent with these tags in #Gene_history_action

Added 'Initial_transfer' as a possible Source in #Evidence.  So when we initially
add Genes with Other_name tags, we can set the Evidence hash with 'Initial_transfer'
to indicate why there is no other Evidence as yet.