Restructure the Evidence hash in the new gene model.  The newly changed part looks like this:
#Evidence  Source  UNIQUE  Author  ?Author
                           Literature   Paper   ?Paper //updated krb 04/03/02
                                        PMID    Text // added for Erich's gene function data 02/03/09
                           Automated_method ?Method ?Database

The PMID tag was all that I wanted to add (for Erich's gene function data) but I then added
a new Literature tag such that PMID and Paper can both have multiple entries.

Still not sure whether UNIQUE should be after Source, wouldn't we want to sometimes attribute
evidence to both a person and a paper?