Reworked the models slightly to better accommodate Erich's gene function data.
This information was to be stored in ?Text fields, this is now being stored in ?LongText
fields which seems much more fitting.

In the Locus class I have now grouped Detailed_description and Provisional_description
under a more general heading 'Gene_information' and moved the current 'Description' tag
under this top level tag.

In the Sequence class, these two fields sit at the bottom of the Visible section.

For the evidence hash, I have appended "_evidence" to all of the tag names, so PMID "Text"
becomes PMID_evidence "Text".  This not only reads better in the tree display, but it should
ensure that if the evidence hash is appended to a class, there won't be any duplication of
tag names (which shouldn't be allowed under acedb) within any class.

Keith