Rejigged to -initial subroutine.
Uncommented some lines which should remove any existing flags.  Still at a halfway
house on this until we sort out our cvs repository.  This stage should automatically
do a cvs update on the database.wrm file but we can not do this now that we have to
have passwords for cvs updates.  Hence this change.