Finally found the annoying source of a very minor error in overlapcheck.
At the same time I got rid of the whole thing and replaced it with overlapcheck.pl
This upgrades to perl5.6.0 and I made a couple of other tiny cosmetic changes.
Also changed post_build_checks.pl to now link to overlapcheck.pl and additionally
made some small changes to the logging information in this script.

Keith