The prepare_build.pl script had sort of been retired from the build guide as some of the things it did were replaced by autoace_minder -initial
However, some of the things it did aren't being done in the current guide, so
I've retired this script and added a stripped down version 'finish_build.pl'
which should be run at the end of the build.

This archives away old wormbase and wormpep releases and updates the current_DB
symlink.

Keith