New checking script.
A script to run at the end of the build.

Takes list of current genomic_canonical clones.

For each clone name it then tries retrieving anything that looks like
a predicted gene, e.g. clone_name.*

For each retrieved object it looks to see if there is a 'Source' tag and
if not it sniffs around finding timestamp information and guessing what tags
from other objects would have created the bogus sequence object.

To be run at the end of the build.  Results are currently emailed to wormbase-dev.

Keith