Added pod documentation to the script.
Forgot to mention that a -d flag allows you to check other databases not just
/wormsrv2/current_DB (which is the default), i.e. check autoace during the build.

ALso, a -v option (verbose) will output errors to the screen as the script runs.
This option also turns off the mailing to wormbase-dev.  So we can use this
to get rid of all the camace bugs before telling wormbase-dev!