New UTR model:
?UTR SMap S_Parent Canonical_parent UNIQUE ?Sequence XREF UTR
      Source_Exons Int UNIQUE Int // start at 1
      Matching_CDS ?Sequence XREF Matching_UTR
      Properties 5'UTR #Evidence
               3'UTR #Evidence
      Species ?Species
      Method ?Method

This is based on what Lincoln asked for to accommodate his UTR data but with some modification
by us (Sanger) to accommodate some changes we would want to make in subsequent releases.
For WS83, 'Source_Exons' and 'Properties' tags will not be used.

Lincoln's intitial dataset creates separate UTR objects for each exon or part of an exon.
Ideally we will switch to using the Source_Exons tags to create just one object for each
5' and 3' UTR.


Intitial set of UTR names are: UTR000001, UTR000002 etc.  This may need to be changed at some point
to allow some more flexibility.  Or could even be dynamically generated during the build to
incorporate current CDS name and include 5' or 3' info.  Up for discussion really.

Keith