New tags in Protein class to allow for logging of changes
Removed the 'Replaced_by' and 'Replaces' tags with:

Wormpep History Int Text ?Text

The Int field will store a Wormpep release number
The first Text field will contain some 'event' information 'E.g. created, removed'.
The second Text field stores the sequence name of a predicted gene.

This change was made to fit in with Anthony's new script which generates history
information for wormpep proteins.  This is a huge improvement on how things were, you can
now see the full history of changes for any wormpep Protein rather than just the name
of what the protein was/became.

This might be expanded a bit more in the future.

Keith