New script to replace something that patch_wormpep used to do.
This script generates a list of new wormpep sequences from the last release
of wormpep, and writes these to a file in the respective wormpep directory
(new_entries.WSxx). This file is needed by the prebuild process.

This replaces the file Analysis_to_do... which was generated by patch_wormpep.

This script is quick and ignores reappeared entries (you could argue that these
too are new proteins but they will already be in the wormprot mysql database).

Keith