Added new subroutine for adding of briggsae assembly data and briggsae gene sets.
This is run by using the -addbriggsae option, very similar to the -addhomol
option.

Also changed script to use perl 5.6.1

Also tried making the log file always have a 'script finished line: date time'.
Some subroutines also add this info, but not for all subroutines and also some
subroutines add this at various times during that subroutine, and so there
should now always be a start and finish time for the autoace_minder
script in the log file.

Also tried adding some more useful information to the email subject line for
all autoace_minder messages. This was previously set to add $logtype to the
subject line, but this variable was always set to "".  Have now asked it
to add $ARGV[0] which should usually be something more useful like "-addhomol"