Many updates to this script.
Primary change is to separate output so that separate emails are sent
to people at CSHL, Sanger, St. Louis and Caltech (separate log files are
also created in addition to one cumulative log file with all possible problems).

Where the source of a problem is unclear or can only be associated to the wormpub
user then these potential errors will go in the Sanger email.

This script additionally looks for valid-looking sequences which do not have a
Corresponding_protein.

Also created a -t (test) flag which runs everything but doesn't send the emails.

Also upgraded to perl5.6.1

This scripts should be run at the end of the build (it is in the build guide but
hasn't always been run).

Keith