Started to add a tag to allow for primary evidence to many classes in the database.
E.g.

?Locus Evidence #Evidence

This is useful for using with data that can come from many sources and where we might
want to see what the primary evidence for the creation of the object was.  E.g. did a
new locus name come from EMBL, Jonathan Hodgkin update email, Andrei curation email,
user comment.

Not all classes really need this though and it is more geared to tracking data which
may be submitted from a user as well as coming in through the standard literature curation.

For now I have added this tag to:

?Locus, ?Allele, ?Gene_class, ?Rearrangement, ?Strain, ?Clone, ?2_pt_data,
?Multi_pt_data, ?Transgene

Please let me know if there are any other classes where you would like to use this
tag with.

These new tags will become available from WS92.

Keith