Lots of small changes
1) Misc. script curation changes (and this script is not even on my list!)

2) Made consistant use of subroutine and variable names to harmonise
discrepancies between mine and Chao-Kung's subroutines

3) Removed a ck1 subroutine that worked out who the user was if in debug mode,
this is all now done by code in the standard script template.  Also allowed
me to remove other sundry bits of code

4) Cleaner log output.  All log errors and warnings now have lines that either
start ERROR: or WARNING: (this is a good practice to get into for all our scripts!)

5) Found out why the number of listed errors were always less than the number reported
by the error counter.  Fixed this.

It's all now shiny and new just ready for Christmas.