EMBLdump.pl had a post-processing routine to add /standard_name qualifiers to CDSs in EMBL records and also use locus names for /gene=... where known.  Also added Product info.
As part of the move to a new transcript class, the acedb internal code now does all
of this for both  CDSs and Transcripts, so there is no need (hopefully) for post
processing.

Script currently points to a separate code installation, when new release of ACEDB comes
out this script can be updated.

Keith