Fixed an old(ish) Dr. Lawson bug which was causing the error messages in make_autoace and also responsible for the very long timestamps that have been appearing in autoace.
Dan's hacking attempted to take a file name of a dumped ace file and
extract just the database name to use as a timestamp user name. E.g.

/wormsrv2/wormbase/briggsae/briggsae_DNA.ace

becomes just 'briggsae'.  But Dan's script actually set -tsuser to be:

/wormsrv2/wormbase/briggsae/briggsae

So when acedb tries to make files in the /database/new directory, it appends
the user name to the file...in this case the -tsuser name makes out that it
contains subdirectories which is what confuses acedb.  So in part this is an
acedb bug.

Got a bug?  I can fix it.