fixed a wide range of errors that were causing annoying -w warning messages. Also found some things that would have never ever worked. Perhaps most important change was making this script handle the models.wrm differently when copying.
It copies this file with -R option to preserve the symbolic link nature...
is this a bad thing?  I think I might have to change this again as other
build scripts may suffer.  Without this fix, the script was ending prematurely
meaning that it never removed the databse .BCK directory.

To discuss tomorrow.  There are still other parts of this script that may not
work, but hey, I can't fix everyones bugs all the time :)