In an attempt to fix one bug (which I did), I seemed to introduce another.
Upgraded one script to perl5.6.1 and this caused AcePerl to give a different
behaviour.  Consider the line of code:

@acc = $obj->DB_info->row();

Under 5.004, the array starts:

$acc[0] = "DB_info"

Under 5.6.1 the array starts:

$acc[0] = "Database".

I.e. the tag name is not now counted as part of the array.

Strange...we really should get everything using 5.6.1 for consistant behaviour