Quite a few changes.
This script was trying to do three things after generating all of the split GFF files.

1) Create new file based on clone_path file to incorporate clone accessions
2) Add wormpep accessions to split gene file
3) Add CDS info to split UTR file

These three stages used to be called by three scripts from within GFFsplitter, but
Anthony recently changed this and incorportated the UTR script into GFFsplitter

I have now:

1) Incorporated all scripts into GFFsplitter and removed the original scripts

2) Fixed anthony's subroutine which wasn't working properly due to a typo

3) Added a check that the UTR subroutine shouldn't run if there isn't any UTR data...
i.e. we run GFFsplitter more than once in the build, but only in the later steps will
there be UTR information in the GFF files that it can split.  So this step now
checks for the B10:Generate_UTR_data lock file.

The new subroutines are all essentially the same and could be cleaned up later on to
remove some redundancy (i.e. one subroutine not three)