GFFsplitter kept on running out of memory when halfway through the last chromosome (getting to the point where it was using over 1Gb).
I *think* that this is because the file handle for each chromosome file wasn't being closed
at the end of each chromsome.  So I've added a simple close statement at the end of each file.

Also, I've added an 'undef(%GFF)' line at the start of each chromosome loop as I think this hash
can safely be reset then.

Keith