Umm, I just needed to make this script also find Transcript->Locus connections in addition to Sequence->Locus connections (to take account of the new Transcript class).
However, I pretty much ended up rewriting the entire script.  It is now faster, and
cleaner with more useful logging information.

The errors that this script flags will probably also be picked up by current_DB_check,
but they are a good source of things for Chao-Kung to check as all of these errors need
to be fixed in geneace.

Keith