Never has so much work been caused by so little a typo! :)
The usage subroutine of autoace_minder will do various things depending on what flag you
pass to it.  If you pass "16" it knows that you haven't run autoace_minder -secure
but you are now trying to run autoace_minder -blat.

However, this is what was in the usage subroutine...


    elsif ($error == 15) {
        # attempted BLAT analysis with agp errors
	...
    }
    elsif ($error == 15) {
        # atempted BLAT analysis without copying to autoace_midway
	...
    }


So there were two 15's instead of a 15 and a 16.  This means that passing error
code 16 to the subroutine doesn't match anything and you will ultimately be
returned to the part of the code that called the usage subroutine.

In the case of the blat routine it means it will try to run autoace_minder -blat
even if you haven't first run autoace_minder -secure.

A case for joint spam award to Dan and Chao-Kung ??? :)

Keith