New subroutine in Wormbase.pm:  &check_write_access
If you pass this a valid database path, it will look for a database/lock.wrm file
in that database. If it exists it returns 'no' as in "no, you can't have write access"...
otherwise it returns "yes".

e.g

$write_access = &check_write_access($database);

if($write_access eq "yes"){
   &save_the_world;
}
else{
   &burn_in_hell;

}


I've modified locus2seq.pl to check for camace write access before trying to
upload the file.  I'm sure there are other places we could/should be using this.

Keith