Two related model changes.
1) New ?GO_code class.  This only contains a couple of fields which can optionally
store the full description of the three-letter GO codes (e.g. IEA) but none of
this is really necessary.


2) Any use of 'GO_term' in any model now also links to ?GO_code.  E.g.

?Some_class
GO_term ?GO_term ?GO_code #Evidence

As a reminder, this allows any one GO term to have one or more GO codes associated
with them, and each GO code can have one or more pieces of evidence.

The Evidence hash was actually missing from a couple of GO_terms in models. wrm so I've
made sure that everything is consistant now.


The GO_code model should be used *EVERYWHERE* that there is a GO_term... this might
mean adding 'ND' (Not Determined) as a GO_code for some GO_terms, but we should in most
cases be able to use the correct GO code.

Keith