added routine to check when a locus is merged into a main name, there should be remark in the gene_class of the main name: This routine deals only with empty remark. Ie. if there is something in the remark, it does not go to "read" to see if remark has been added.
Output sample:

vab-21 became an other_name of ifb-1 and no remark is added in Gene_class ifb
VAMP/1 became an other_name of snb-1 and no remark is added in Gene_class snb

Total of 470 are picked up and corrected by ace file. There are 243 to check by hand at this point.