New script which simply emails a list of checks to the next person who will be doing the build.  E.g. in this build dan should run:
next_builder_checks.pl -user krb

This sends me an email and the idea is that I only hand back to Dan when I've
covered all of these things.

*IMPORTANT*
Each time we discover new problems we should add to this script to spot them in
the future!

I'll now update the build guide

Keith