Two main changes.
The first and more serious is that new proteins that were classified
as 'reappeared' genes in the wormpep.diff file were not getting added
to the file of new proteins and hence not being analysed on the farm
with BLAST etc.

This problem reflects differences in how people treat 'reappeared' genes.

The script now will add any new protein sequence to this file as long as
the CE number didn't appear in the previous wormpep release wp.fasta file.

The other change I made makes sure that genes that are flagged as 'changed' or
'new' but have a peptide that already exists, also don't get dumped to the
file.  This is not so serious as Anthony has written a downstream check in
another script that removes any duplicate sequences before adding to the
MySQL database.

Keith