I was attempting to debug this script to find out why it wasn't runnning on the cronjob.  This turned out to be a data issue, but I also found out some other bugs...
split camcheck log files were not including the pad,ar2,dl1 name as intended.  This is
now fixed.

Also made -d (debug) accept a user name and only email the report to that user (in
keeping with the behaviour of other scripts).  Was previously hard-coded to only email
ar2

Also made a new option -v for verbose (to turn on extra output to screen).  This was
previously done under the -d option.  Again this is in keeping with other scripts.

Keiht