New script which is basically a complete rewrite of what was list_loci_designations.
This script now uses about 1/3 of the code of the old script and runs about 100
times faster.

This script has been divorced from the build process.  It will now run against
geneace and publish the latest set of gene names (and sequence connections and other
names) every day.  The output of this is now properly hyperlinked to WormBase.

Plus, the script manages to make the webpublish thing work which is an added bonus.
Whether this will still work under a crontab job is another matter!

Keith