Generates UTRs with splicing. The splicing is based on introns confirmed_in_UTR that fall within the span of matching cDNAs attached to a gene. Therefore relies on matching_cDNAs to be set correctly which in some cases is not the case.
Known problems:
If a gene starts immediately after a UTR intron then there is no UTR exon pair to span the intron.

A UTR spans a confirmed intron, but the cDNAs associated with the CDS do not have this splicing ie variable UTR splicing.  The intron is incorporated in to the UTR contrary to what the cDNAs indicate.

Apart from that it works OK :)