Sorry but I couldn't help myself...make_acefiles is now make_acefiles.pl and referenced as such in autoace_minder.
The main change I made was to make this script check that an ace file has been created
for each line in autoace_config.  If there is no data in that corresponding class in
acedb, no output file gets dumped from tace, and so downstream processing steps of that
file are meaningless.  The script now skips downstream processing steps if no file output
was created.

Now to test it...