Upgraded to perl 5.8.0 but more importantly
-check is no longer an option.  This used to make/check agp files *and* run dbcomp.pl.
This is now handled by the following two options:

-agp
-dbcomp

The reason for this is that we are now going to start running -agp twice.  Once early in the
build (as usual) and then again right at the end of the build.  Any EMBL synchronisation
issues that cause errors the first time should be fixed by the second time as EMBL would have
had a week to synchronise things.

Also done some misc. tidying up removing unused options.

Keith