THE BIG CHANGE!!!
Despite changing countless lines of code in this, blat_them_all.pl and autoace_minder,
this is the big one.  The couple of characters change will mean:

blat scores (for purpose of deciding BEST vs OTHER) will be calculated by dividing
the blat output 'match' score by the total length of the the query sequence and NOT
the sum of the matching block sizes.

The current scenario meant that 30 bp chunks of an mRNA sequence (usually corresponding
to a polyA tail) were getting scores of 100 despite only representing 1% of the parent
mRNA.  This was because in the blat output, the sum of the lengths of the blocks nearly
always (unless there is a mismatch) equal the score from 'match' column...so most short
matches end up with a 100% score, whereas an mRNA match of 2000 bp which has 4 mismatches
would get a lower score.

This change *should* mean that lots of small, shitty things which are currently BEST
will become OTHER and lots of good, long matches, which are currently OTHER will become
BEST.

We will see!!!

Keith