Changed all use of:
$runtime = &runtime; print LOG "Something happened at $runtime\n";

to:

print LOG "Something happened at ",&runtime,"\n";

Which is a bit cleaner with the code.  Also started to overhaul the use of
ERROR lock files in /wormsrv2/autoace/logs

This is ongoing, but I want to add more stops so that you can't proceed
with step 2 if there were errors in step 1.  You will first have to remove a lock
file by hand.

Also I will start adding tests for all system calls, e.g.

my $status = system(....);

if ($status != 0){

   print LOG "ERRROS";

}

There are still lots of scripts that are failing and they just cause autoace_minder to stop
without emailing the log file because of the && die comments.

Keith