First set of model changes for WS114
These are the changes for acedb to work with Ed's new code (partly relating
to new gene model).

i) New 'Properties Allele' line in ?Allele class, all alleles must use this else
they won't be dumped as Alleles in GFF.  This tag ('Allele') is a magic tag.

ii) New 'Properties' prefix to 'Homol' tag.  'Homol' is a magic tag too, so this
is for consistency with the above

iii) New lines in ?Transposon, and ?Pseudogene ('Properties Transposon' and
'Properties Pseudogene').

For now we don't need these and they are not being used as magic tags by acedb.  But we
also have 'Properties Transcript' in the ?Transcript class which is used (though is not
a magic tag) and have 'Properties CDS' in the ?Sequence class ('CDS' is a magic tag and
this will move to the new ?CDS class). So I've added these such that we can start using
them if we want to be completely consistant for all our SMapped classes which are gene-like
(?Sequence,?CDS,?Pseudogene,?Transcript, and ?Transposon).  It may be that we need to use
these in future if we ever need to make acedb treat these classes in functionally different
ways.