Added new line to Evidence hash:
CGC_disapproved Text #Evidence

Can be used to flag locus/allele names that are in the database but which were
never approved by the CGC and which constantly cause confusion.  Note that the
names in question might be valid but a user first used it without permission.  I.e.

Locus : dmo-1
Other_name "mod-1" CGC_disapproved "mod-1 has been used to refer to this gene but mod-1
is actually a different gene altogether" Paper_evidence "[cgc3363]"

Note that for most of the ~25 cases in the Locus class where this can be used, we have
already added Remarks to explain the mix up but this helps make this a bit more formal.